타임특가, 리뷰 답변 추가 구현

+++ back/src/main/java/com/admin/back/repository/ReviewRepository.java
@@ -4,5 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.admin.back.entity.ReviewEntity;
 
-public interface ReviewRepository extends JpaRepository<ReviewEntity, Integer> {
+public interface ReviewRepository extends JpaRepository<ReviewEntity, Long> {
+    
 }
